last version

+++ src/SalesManager.java
@@ -15,9 +15,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = sales[0];
-        for (int sale: sales) {
+    public long min() {
+        long min = sales[0];
+        for (long sale: sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -27,7 +27,7 @@ public class SalesManager {
 
     public int mean() {
         int sum = 0;
-        for (int sale: sales) {
+        for (long sale: sales) {
             sum += sale;
         }
         sum -= this.min();
